+++ axelor-account/src/main/java/com/axelor/apps/account/service/AccountingReportService.java
@@ -67,4 +67,7 @@ public interface AccountingReportService {
   public BigDecimal getCreditBalance(AccountingReport accountingReport, String queryFilter);
 
   public BigDecimal getCreditBalanceType4();
+
+  public String getReportFileLink(AccountingReport accountingReport, String name)
+      throws AxelorException;
 }