+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/report/IReport.java
@@ -23,4 +23,5 @@ public interface IReport {
   public static final String EBICS_CERTIFICATE = "EbicsCertificate.rptdesign";
   public static final String BANK_STATEMENT_AFB120 = "BankStatementAFB120.rptdesign";
   public static final String BANK_RECONCILIATION = "BankReconciliation.rptdesign";
+  public static final String BANK_PAYMENT_REPORT_TYPE = "BankReconciliationStatement.rptdesign";
 }